Add all logic for the user dashboard
In UserActivity.java initialize the Spinner to get the questions from
the database. checkSelectedVote() function decides which fragment to
show (UserVotesFragment, UserResultsFragment or UserNoResultsFragment). Added functions to check if vote is ended.

In UserVotesFragment.java initialize the view and set Listener on the
submit button.

In UserResultsFragment.java calculate the votes and display them.